Move captured_output function tests global package
