+++ tests/install_queue/test_basic_event.py
@@ -2,14 +2,14 @@ import unittest
 
 from egg.install_queue.install_event import InstallEvent
 from egg.install_queue.install_queue import InstallQueue
-from tests.install_queue import captured_output
+from tests import captured_output
 
 
 class BasicTestClass(object):
     def __init__(self, method_name: str):
         self.method_name = method_name
 
-    def printTest(self, value: int):
+    def print_test(self, value: int):
         print('method_name: %s, value: %s' % (self.method_name, value))
 
 
@@ -26,9 +26,9 @@ class BasicEvent(InstallEvent):
 class TestBasicEvent(unittest.TestCase):
 
     def test_basic_event(self):
-        event = BasicEvent(BasicTestClass.printTest.__name__, value=1)
+        event = BasicEvent(BasicTestClass.print_test.__name__, value=1)
         InstallQueue().add(event)
         with captured_output() as (out, err):
             InstallQueue().execAll()
         output = out.getvalue().strip()
-        self.assertEqual(output, 'method_name: printTest, value: 1')
+        self.assertEqual(output, 'method_name: print_test, value: 1')
